PR-202

+++ module/userProjects/helper.js
@@ -119,7 +119,9 @@ module.exports = class UserProjectsHelper {
                     "entityInformation._id",
                     "lastDownloadedAt",
                     "appInformation",
-                    "status"
+                    "status",
+                    "programId",
+                    "isAPrivateProgram"
                 ]);
                 
                 if (!userProject.length > 0) {
@@ -158,6 +160,19 @@ module.exports = class UserProjectsHelper {
                 }
                 let createNewProgramAndSolution = false;
                 let solutionExists = false;
+                
+                if (
+                    (userProject[0].programId && !data.programId) ||
+                    (userProject[0].solutionInformation?._id && !data.solutionId)
+                ) {
+                    throw {
+                        status: HTTP_STATUS_CODE['bad_request'].status,
+                        message:
+                        userProject[0].programId && !data.programId
+                                ? CONSTANTS.apiResponses.REQUIRED_PROGRAM_ID
+                                : CONSTANTS.apiResponses.REQUIRED_SOLUTION_ID
+                    };
+                }
 
                 if (data.programId && data.programId !== "") {
 
@@ -229,7 +244,7 @@ module.exports = class UserProjectsHelper {
                         return resolve(programAndSolutionInformation);
                     }
 
-                    if (solutionExists) {
+                    if (solutionExists && userProject[0].isAPrivateProgram) {
 
                         let updateProgram =
                             await surveyService.removeSolutionsFromProgram(
@@ -361,6 +376,15 @@ module.exports = class UserProjectsHelper {
                 if ( data.status == CONSTANTS.common.COMPLETED_STATUS || data.status == CONSTANTS.common.SUBMITTED_STATUS ) {
                     updateProject.completedDate = new Date();
                 }
+
+                // remove restricted keys
+                updateProject = _.omit(updateProject, [
+                                          "programInformation",
+                                          "solutionInformation",
+                                          "programId",
+                                          "solutionId"
+                                        ]);
+                
                 
                 let projectUpdated =
                     await projectQueries.findOneAndUpdate(
@@ -2544,8 +2568,7 @@ module.exports = class UserProjectsHelper {
                     recipient : {
                         id : data.userId,
                         name : certificateUserName,
-                        type : data.userProfile.profileUserType.type,
-                        userLocations : data.userProfile.userLocations
+                        type : data.userProfile.profileUserType.type
                     },
                     templateUrl : data.certificate.templateUrl,
                     issuer : certificateTemplateDetails[0].issuer,
